Clean useless stuff
