+++ packages/twenty-server/src/engine/api/rest/core/controllers/rest-api-core.controller.ts
@@ -13,8 +13,7 @@ import {
 
 import { Request, Response } from 'express';
 
-import { RestApiCoreServiceV2 } from 'src/engine/api/rest/core/rest-api-core-v2.service';
-import { RestApiCoreService } from 'src/engine/api/rest/core/rest-api-core.service';
+import { RestApiCoreService } from 'src/engine/api/rest/core/services/rest-api-core.service';
 import { RestApiExceptionFilter } from 'src/engine/api/rest/rest-api-exception.filter';
 import { JwtAuthGuard } from 'src/engine/guards/jwt-auth.guard';
 import { WorkspaceAuthGuard } from 'src/engine/guards/workspace-auth.guard';
@@ -23,49 +22,46 @@ import { WorkspaceAuthGuard } from 'src/engine/guards/workspace-auth.guard';
 @UseGuards(JwtAuthGuard, WorkspaceAuthGuard)
 @UseFilters(RestApiExceptionFilter)
 export class RestApiCoreController {
-  constructor(
-    private readonly restApiCoreService: RestApiCoreService,
-    private readonly restApiCoreServiceV2: RestApiCoreServiceV2,
-  ) {}
+  constructor(private readonly restApiCoreService: RestApiCoreService) {}
 
   @Post('batch/*')
   async handleApiPostBatch(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.createMany(request);
+    const result = await this.restApiCoreService.createMany(request);
 
     res.status(201).send(result);
   }
 
   @Post('*/duplicates')
   async handleApiFindDuplicates(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.findDuplicates(request);
+    const result = await this.restApiCoreService.findDuplicates(request);
 
     res.status(200).send(result);
   }
 
   @Post('*')
   async handleApiPost(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.createOne(request);
+    const result = await this.restApiCoreService.createOne(request);
 
     res.status(201).send(result);
   }
 
   @Get('*')
   async handleApiGet(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.get(request);
+    const result = await this.restApiCoreService.get(request);
 
     res.status(200).send(result);
   }
 
   @Delete('*')
   async handleApiDelete(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.delete(request);
+    const result = await this.restApiCoreService.delete(request);
 
     res.status(200).send(result);
   }
 
   @Patch('*')
   async handleApiPatch(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.update(request);
+    const result = await this.restApiCoreService.update(request);
 
     res.status(200).send(result);
   }
@@ -75,7 +71,7 @@ export class RestApiCoreController {
   // of PATCH, and because the PUT verb is often used as a PATCH.
   @Put('*')
   async handleApiPut(@Req() request: Request, @Res() res: Response) {
-    const result = await this.restApiCoreServiceV2.update(request);
+    const result = await this.restApiCoreService.update(request);
 
     res.status(200).send(result);
   }
