+++ packages/twenty-server/src/engine/api/rest/core/rest-api-core.module.ts
@@ -4,8 +4,8 @@ import { HttpModule } from '@nestjs/axios';
 import { RestApiDeleteOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-delete-one.handler';
 import { RestApiCreateOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-create-one.handler';
 import { RestApiUpdateOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-update-one.handler';
-import { RestApiGetOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-get-one.handler';
-import { RestApiGetManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-get-many.handler';
+import { RestApiFindOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-find-one.handler';
+import { RestApiFindManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-find-many.handler';
 import { CoreQueryBuilderModule } from 'src/engine/api/rest/core/query-builder/core-query-builder.module';
 import { WorkspaceCacheStorageModule } from 'src/engine/workspace-cache-storage/workspace-cache-storage.module';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
@@ -13,8 +13,7 @@ import { RecordTransformerModule } from 'src/engine/core-modules/record-transfor
 import { WorkspacePermissionsCacheModule } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.module';
 import { RestApiCoreController } from 'src/engine/api/rest/core/controllers/rest-api-core.controller';
 import { coreQueryBuilderFactories } from 'src/engine/api/rest/core/query-builder/factories/factories';
-import { RestApiCoreServiceV2 } from 'src/engine/api/rest/core/rest-api-core-v2.service';
-import { RestApiCoreService } from 'src/engine/api/rest/core/rest-api-core.service';
+import { RestApiCoreService } from 'src/engine/api/rest/core/services/rest-api-core.service';
 import { RestApiService } from 'src/engine/api/rest/rest-api.service';
 import { ApiEventEmitterService } from 'src/engine/api/graphql/graphql-query-runner/services/api-event-emitter.service';
 import { AuthModule } from 'src/engine/core-modules/auth/auth.module';
@@ -26,8 +25,8 @@ const restApiCoreResolvers = [
   RestApiCreateOneHandler,
   RestApiCreateManyHandler,
   RestApiUpdateOneHandler,
-  RestApiGetOneHandler,
-  RestApiGetManyHandler,
+  RestApiFindOneHandler,
+  RestApiFindManyHandler,
   RestApiFindDuplicatesHandler,
 ];
 
@@ -45,7 +44,6 @@ const restApiCoreResolvers = [
   providers: [
     RestApiService,
     RestApiCoreService,
-    RestApiCoreServiceV2,
     ApiEventEmitterService,
     ...coreQueryBuilderFactories,
     ...restApiCoreResolvers,
