+++ packages/twenty-server/src/engine/api/rest/core/services/rest-api-core.service.ts
@@ -7,20 +7,20 @@ import { parseCorePath } from 'src/engine/api/rest/core/query-builder/utils/path
 import { RestApiDeleteOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-delete-one.handler';
 import { RestApiCreateOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-create-one.handler';
 import { RestApiUpdateOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-update-one.handler';
-import { RestApiGetOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-get-one.handler';
-import { RestApiGetManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-get-many.handler';
+import { RestApiFindOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-find-one.handler';
+import { RestApiFindManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-find-many.handler';
 import { RestApiCreateManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-create-many.handler';
 import { RestApiFindDuplicatesHandler } from 'src/engine/api/rest/core/handlers/rest-api-find-duplicates.handler';
 
 @Injectable()
-export class RestApiCoreServiceV2 {
+export class RestApiCoreService {
   constructor(
     private readonly restApiDeleteOneHandler: RestApiDeleteOneHandler,
     private readonly restApiCreateOneHandler: RestApiCreateOneHandler,
     private readonly restApiCreateManyHandler: RestApiCreateManyHandler,
     private readonly restApiUpdateOneHandler: RestApiUpdateOneHandler,
-    private readonly restApiGetOneHandler: RestApiGetOneHandler,
-    private readonly restApiGetManyHandler: RestApiGetManyHandler,
+    private readonly restApiFindOneHandler: RestApiFindOneHandler,
+    private readonly restApiFindManyHandler: RestApiFindManyHandler,
     private readonly restApiFindDuplicatesHandler: RestApiFindDuplicatesHandler,
   ) {}
 
@@ -48,9 +48,9 @@ export class RestApiCoreServiceV2 {
     const { id: recordId } = parseCorePath(request);
 
     if (isDefined(recordId)) {
-      return await this.restApiGetOneHandler.handle(request);
+      return await this.restApiFindOneHandler.handle(request);
     } else {
-      return await this.restApiGetManyHandler.handle(request);
+      return await this.restApiFindManyHandler.handle(request);
     }
   }
 }
